Reorganize somewhat.
