+++ lib/raml/node/abstract_method.rb
