+++ lib/raml/node/abstract_resource.rb
