+++ lib/raml/node/body.rb
