+++ lib/raml/node/documentation.rb
