+++ lib/raml/node/header.rb
