+++ lib/raml/node/method.rb
