+++ lib/raml/node/parameter/abstract_parameter.rb
