+++ lib/raml/node/parameter/base_uri_parameter.rb
