+++ lib/raml/node/parameter/query_parameter.rb
