+++ lib/raml/node/parameter/uri_parameter.rb
