+++ lib/raml/node/protocol.rb
