+++ lib/raml/node/resource.rb
