+++ lib/raml/node/resource_type.rb
