+++ lib/raml/node/resource_type_reference.rb
