+++ lib/raml/node/response.rb
