+++ lib/raml/node/root.rb
