+++ lib/raml/node/schema.rb
