+++ lib/raml/node/schema_reference.rb
