+++ lib/raml/node/template.rb
