+++ lib/raml/node/trait.rb
