+++ lib/raml/node/trait_reference.rb
