+++ lib/raml/patch/hash.rb
