+++ lib/raml/patch/module.rb
